Carla.CPM — Type

CPM is the data structure for Carla Probability Model.

Keword Arguments

emissionprob – Emission Probability function, type of ResponseFunction

transitionprob – Transition Probability function, type of ResponseFunction

initialvwec – Initial Weight Vector, Vector of Float64

varianceprior – Variance Prior, a Float64

initparamstd – Initial Parameter S.D., a Float64

estimatebeta – Estimate Beta Parameters?, Bool type

estimatedelta – Estimate Delta Parameters?, Bool type

Notes

All the fields can be assigned using the appropriate keywords. The initialization constructor CPM(), uses the values in the example, as default for the fields in CPM.

Example

M1 = CPM( emissionprob=DINA(), transitionprob=DINA(), initialwvec = [1.2, 0.6], varianceprior = 100/536, initparamstd = 0.2, estimatebeta = true, estimatedelta = false)

Carla.DINA — Type

DINA <: ResponseFunction

DINA Response Function

Carla.DINO — Type

DINO <: ResponseFunction

DINO Response Function

Carla.FUZZYDINA — Type

FUZZYDINA <: ResponseFunction

FUZZYDINA Response Function

Carla.ResponseFunction — Type

ResponseFunction

Abstract type for response functions