Carla.emission_responsevec — Method

emission_responsevec(
    EmissionType::ResponseFunction, QMatrix::Matrix,
    itemID::Integer, t::Integer, αMatrix::Matrix
)

Computes the local emission response vectors $\psi_{jt} (\alpha)$

Arguments

EmissionType: Emission Response Type (DINA(),DINO(), or FUZZYDINA())

QMatrix: J (No. of Items) × K (No. of Skills) Matirx

itemID: Integer denoting the jth item

t: time index

αMatrix: K (No. of Skills) × T (No. of Timepoints) Matrix

Output:

A 2-element Vector{Float64} (Probability Vectors)

Note: This is equivalent to the psiemission function in MATLAB Carla

Carla.transition_responsevec — Method

transition_responsevec(
    EmissionType::ResponseFunction, RMatrix::Matrix,
    skillID::Integer, t::Integer, αMatrix::Matrix
)

Computes the local transition response vectors $\phi_{kt} (\alpha)$

Arguments

EmissionType: Emission Response Type (DINA(),DINO(), or FUZZYDINA())

RMatrix: K (No. of Skills) × K (No. of Skills) Matirx

skillID: Integer denoting the jth skill

t: time index

αMatrix: K (No. of Skills) × T (No. of Timepoints) Matrix

Output:

A 2-element Vector{Float64} (Probability Vectors)

Note: This is equivalent to the phitransition function in MATLAB Carla